Fixed changes for RockSample; everything seems to work!

+++ src/problems/rocksample/RockSampleModel.hpp
@@ -45,10 +45,10 @@ class RockSampleObservation;
 
 struct RockSampleChange : public solver::ModelChange {
     std::string changeType = "";
-    double i0 = 0;
-    double i1 = 0;
-    double j0 = 0;
-    double j1 = 0;
+    long i0 = 0;
+    long i1 = 0;
+    long j0 = 0;
+    long j1 = 0;
 };
 
 class RockSampleModel : public ModelWithProgramOptions {
